ValidParentheses/ValidParenthesesJava/src/main/java/com/leetcode/validparentheses/Solution.java
package com.leetcode.validparentheses;

import java.util.LinkedList;
import java.util.NoSuchElementException;

public class Solution {
    // initial solution
    public boolean isValid1(String s) {
        LinkedList<Character> stack = new LinkedList<>();
        stack.push(s.charAt(0));
        for (int i = 1; i < s.length(); i++) {
            char c = s.charAt(i);
            try {
                switch (c) {
                    case ')':
                        if (stack.pop() != '(') {
                            return false;
                        }
                        break;
                    case '}':
                        if (stack.pop() != '{') {
                            return false;
                        }
                        break;
                    case ']':
                        if (stack.pop() != '[') {
                            return false;
                        }
                        break;
                    default:
                        stack.push(c);
                }
            } catch (NoSuchElementException e) {
                return false;
            }
        }
        return stack.isEmpty();
    }

    // cleaner solution
    public boolean isValid(String s) {
        LinkedList<Character> stack = new LinkedList<>();
        for (int i = 0; i < s.length(); i++) {
            char c = s.charAt(i);
            switch (c) {
                case '(':
                    stack.push(')');
                    break;
                case '{':
                    stack.push('}');
                    break;
                case '[':
                    stack.push(']');
                    break;
                default:
                    if(stack.isEmpty() || stack.pop() != c) {
                        return false;
                    }
            }
        }
        return stack.isEmpty();
    }

}
